apps/worker/requirements.txt
celery==5.3.0
redis==5.0.0
pydantic==2.5.0
requests==2.31.0
Pillow==10.3.0
PyMuPDF==1.23.0
python-pptx==0.6.21
gTTS==2.4.0
openai==1.0.0
anthropic==0.8.0
google-generativeai==0.3.2
python-docx==1.1.0
openpyxl==3.1.2
openai-whisper==20231117
whisperx==3.1.1
